requirements.txt
pyautogui
pyperclip
beautifulsoup4
python-dotenv
pyobjc-core # pyautogui dependency
pyobjc  # pyautogui dependency
